Implementado tratamento de exceções

+++ src/main/java/com/jcsoftware/orders/services/UserService.java
@@ -4,11 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.jcsoftware.orders.entities.User;
 import com.jcsoftware.orders.repositories.UserRepository;
+import com.jcsoftware.orders.services.exceptions.IntegrityViolationException;
 import com.jcsoftware.orders.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -32,15 +36,33 @@ public class UserService {
     }
     
     public void delete(Long id) {
-    	 repository.deleteById(id);
+    	
+    	try {
+    		if(repository.existsById(id)) {
+    			repository.deleteById(id);	
+    		} else {
+    			throw(new ResourceNotFoundException(id));
+    		}
+    			
+    	}
+    	
+    	catch(DataIntegrityViolationException e) {
+    		throw(new IntegrityViolationException(id));
+    	}
+    	 
     }
     
     public User update(Long id,User obj) {
     	
-     User entity = repository.getReferenceById(id);
-     updateData(entity,obj);
-     return repository.save(entity);
-     
+	    try {	
+	     User entity = repository.getReferenceById(id);
+	     updateData(entity,obj);
+	     return repository.save(entity);
+	    } 
+	    catch (EntityNotFoundException e) {
+	    	throw(new ResourceNotFoundException(id));
+	    }
+    
     }
     
     private void updateData(User entity, User obj) {
